requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-ipware==7.0.1
django-jalali-date==1.1.3
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jalali_core==1.0.0
jdatetime==5.0.0
kavenegar==1.1.2
openpyxl==3.1.5
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
requests==2.32.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
zope.interface==7.1.0
